add utils to chemnlp module
